LF-13165 imx91_evk/qsb: Align default mmcroot env with mmcdev
On iMX91, default mmcdev is set to 0 (eMMC) for flexspi NAND boot.
But default mmcroot was not changed. There is no issue because we
have enabled dynamic mmc env setting, so mmcroot will be updated to
align with mmcdev at runtime.
However, it is better to set same default values for both mmcdev
and mmcroot.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>