MA-22915 Move ele buffer address to avoid Kernel image be overwrite
For imx95, ele has no permission to access heap or stack area, so a memory area
in DRAM be set as ele buffer to exchange data.
As before, we set CONFIG_ELE_SHARED_BUFFER=0x92800000. This address has possibility
that conflict with Kernel load address.
This patch modify the ele buffer as '0x98000000' to avoid any overwrite risk.

Signed-off-by: Maximus <maximus.sun@nxp.com>