LFU-812 arm: dts: imx95: Assign PLL VCO as the parent of PLL
U-boot SCMI clock driver does not setup parent relationship for clocks, because
the "get possible clock parent" is not implemented. The latest SM has changed to
not implicitly enable PLL VCO when PLL clock is enabled. So we have to enable
PLL VCO explicitly in u-boot.
The fix is assigning VIDEO PLL VCO as VIDEO PLL parent and assiging LDB PLL VCO
as LDB PLL parent to resolve the problem.

Signed-off-by: Ye Li <ye.li@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>